+++ src/Modules/Informative/Public/components/Hero.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import type { HeroSection } from '../services/informativeService';
+import type { HeroSection } from '../../services/informativeService';
 import '../styles/public-view.css';
 
 interface HeroProps {
